+++ Lib/TSCompiler/ProjectOptions.cs
@@ -50,7 +50,6 @@ namespace Lib.TSCompiler
         public bool NoHtml;
         public bool WarningsAsErrors;
         public bool PreserveProjectRoot;
-        public bool ForceNewBundler;
         public string JasmineVersion;
         public List<string> TestDirectories;
         public string PathToTranslations;
@@ -508,7 +507,6 @@ namespace Lib.TSCompiler
                 ProcessEnvs!["NODE_ENV"] = Parser.Parse("DEBUG?\"development\":\"production\"");
             }
             PreserveProjectRoot = bobrilSection?["preserveProjectRoot"]?.Value<bool>() ?? false;
-            ForceNewBundler = bobrilSection?["forceNewBundler"]?.Value<bool>() ?? false;
             ProxyUrl = GetStringProperty(bobrilSection, "proxyUrl", null);
         }