+++ Lib/Utils/CommandLineParser/Definitions/BuildCommand.cs
@@ -15,7 +15,7 @@ namespace Lib.Utils.CommandLineParser.Definitions
             new CommandLineArgumentBool(description: "quick debuggable bundling", words: new[] {"-f", "--fast"});
 
         public CommandLineArgumentBool NewBundler { get; } = new CommandLineArgumentBool(
-            description: "use new Njsast based bundler", words: new[] {"-x", "--newBundler"}, defaultValue: false);
+            description: "use new Njsast based bundler", words: new[] {"-x", "--newBundler"}, defaultValue: true);
 
         public CommandLineArgumentEnumValues SourceMap { get; } =
             new CommandLineArgumentEnumValues("build sourcemaps", new[] {"-g", "--sourcemap"}, new[] {"no", "yes"});